working on htaccess
learning htaccess and changing the url as required for SEO solution

+++ downloadZip.php
@@ -1,16 +1,25 @@
 <?php
 
-// name of the zip file
-$yourfile = $_GET['name'] . ' Album Pics' . '.zip';
-// code for downloading the zip file from the server
-header("Content-Type: application/zip");
-header("Content-Disposition: attachment; filename=" . $yourfile . "");
-header("Content-Length: " . filesize($yourfile));
-readfile($yourfile);
-
-if (file_exists($yourfile)) {
-    // deleteing the zip file after downloading it
-    unlink($yourfile);
+$access = 'my_value';
+if (empty($access)) {
+    header("Location: index");
+    die();
 }
-exit;
+
+if (isset($_GET['name'])) {
+// or however you get the path
+    $yourfile = $_GET['name'] . ' Album Pics' . '.zip';
+
+    header("Content-Type: application/zip");
+    header("Content-Disposition: attachment; filename=" . $yourfile . "");
+    header("Content-Length: " . filesize($yourfile));
+    readfile($yourfile);
+
+    if (file_exists($yourfile)) {
+        unlink($yourfile);
+    }
+    exit;
+} else {
+    header('Location: ./');
+}
 ?>
